+++ packages/common/CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
-## Unreleased
+## 5.15.1 ((10/3/2023, 06:57 PM PST))
+
+This is an artificial version bump with no new change.
 
 #### 📘 Misc